// HFY.MvcWebUI.Angular/ClientApp/src/app/lazy-general-page/classes/lastikler/models/lastik-guncelle.ts
export class LastikGuncelle {
    LastikID: number;


    FirmaID: number;

    LastikMarkaID: number;

    LastikMarkaDesenID: number;

    SeriNo: string;

    Fiyat: number;

    EbatID: number;

    LastikTurID: number;

    LastikTipID: number;

    DisSeviyesi: number;

    LastikKilometre: number;

    LastikKonumID: number;

    ListeAktiflik: boolean;
}